+++ internal/hitrecord/hitrecord.go
@@ -1,17 +1,17 @@
 package hitrecord
 
-import "github.com/flynn-nrg/izpi/internal/vec3"
+import "github.com/flynn-nrg/go-vfx/math32/vec3"
 
 // HitRecord contains data related to an intersection between a ray and an object.
 type HitRecord struct {
-	u      float64
-	v      float64
-	t      float64
+	u      float32
+	v      float32
+	t      float32
 	p      vec3.Vec3Impl
 	normal vec3.Vec3Impl
 }
 
-func New(t float64, u float64, v float64, p vec3.Vec3Impl, normal vec3.Vec3Impl) *HitRecord {
+func New(t float32, u float32, v float32, p vec3.Vec3Impl, normal vec3.Vec3Impl) *HitRecord {
 	return &HitRecord{
 		u:      u,
 		v:      v,
@@ -32,16 +32,16 @@ func (hr *HitRecord) P() vec3.Vec3Impl {
 }
 
 // T returns the t value.
-func (hr *HitRecord) T() float64 {
+func (hr *HitRecord) T() float32 {
 	return hr.t
 }
 
 // U returns the u value.
-func (hr *HitRecord) U() float64 {
+func (hr *HitRecord) U() float32 {
 	return hr.u
 }
 
 // V returns the V value.
-func (hr *HitRecord) V() float64 {
+func (hr *HitRecord) V() float32 {
 	return hr.v
 }
